% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olapwrt.R
\name{addgraph}
\alias{addgraph}
\title{addgraph}
\usage{
addgraph(repname, foldid, g, currfilt, M)
}
\arguments{
\item{repname}{is name of report being saved}

\item{foldid}{is the folder id under which this report is being saved}

\item{g}{is the graph being saved}

\item{currfilt}{are the filters in use}

\item{M}{is the metadata/config database connection}
}
\description{
addgraph saves a new graph into the databsae
}
\details{
add graph_series

add graph_def

add folder_tree

add graph_props
}
