% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodata.R
\name{dodata}
\alias{dodata}
\title{dodata}
\usage{
dodata(M, D, measures, dims, filters, gtype)
}
\arguments{
\item{M}{is database connection for Meta data}

\item{D}{is database connection fro data}

\item{measures}{are measures selected}

\item{dims}{are dimensions selected}

\item{filters}{are filters selected}

\item{gtype}{is graph type}
}
\description{
create a dataframe that contain rows from selected measures, dimensions and filters
}
