% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dograph.R
\name{dograph}
\alias{dograph}
\title{dograph}
\usage{
dograph(M, D, measures, dims, filters, gtype)
}
\arguments{
\item{M}{is database connection for Meta data}

\item{D}{is database connection fro data}

\item{measures}{are measures selected}

\item{dims}{are dimensions selected}

\item{filters}{are filters selected}

\item{gtype}{is graph type}
}
\description{
create an aggregate dataframe from selected measures, dimensions and filters
}
