% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxy.R
\name{getgraph}
\alias{getgraph}
\title{getgraph}
\usage{
getgraph(cfg, id)
}
\arguments{
\item{cfg}{is the meta data data base}

\item{id}{is the report id}
}
\description{
importFrom rlang .data
}
\details{
Returns graph structure for a report id
}
