% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxy.R
\name{getxy}
\alias{getxy}
\title{getxy}
\usage{
getxy(cfg, id)
}
\arguments{
\item{cfg}{is the database connection for meta data (config) database}

\item{id}{is report id}
}
\description{
look up and return a saved report
}
