% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savefilter.R
\name{savefilter}
\alias{savefilter}
\title{savefilter}
\usage{
savefilter(my_cfg, uid, fil, fname)
}
\arguments{
\item{my_cfg}{is meta data/config database connection}

\item{uid}{is the logged in user id}

\item{fil}{is the filter detail}

\item{fname}{is the name of the filter to be saved}
}
\description{
save a filter into database
}
\details{
filter_header and filter_details tables are written
}
