% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgraph.R
\name{setrepid}
\alias{setrepid}
\title{setrepid}
\usage{
setrepid(cfg, repid)
}
\arguments{
\item{cfg}{is config/meta data database connection}

\item{repid}{is report id to be retrieved}
}
\description{
retrieve a saved report into the graph structure
}
\details{
returns the graph structure that stores all the graph properties retrived for a report
}
